feat: enhance Map component with direct zone and point fetching, and remove unused hooks
